refactor: split main.rs into cli/ module tree (#113)
Extract CLI handler functions from main.rs (3763 lines) into focused
modules under src/cli/:

- cli/mod.rs: shared utilities (parse_skill_ref, print_safety_report)
- cli/author.rs: validate, pack, publish, init-skill, init-registry
- cli/install.rs: install command
- cli/search.rs: search, categories, info, list
- cli/trust.rs: trust subcommands and audit
- cli/setup.rs: setup command

main.rs retains: entry point, arg structs, Command enum, MCP server
wiring, and the test module. Reduced from 3763 to 2486 lines.

Pure mechanical move -- no logic changes.

Closes #113
