+++ src/main.rs
@@ -3,6 +3,7 @@
 //! Binary entry point. CLI parsing (clap), MCP server setup (tower-mcp),
 //! and transport management. Core logic lives in the library crate.
 
+mod cli;
 mod resources;
 mod tools;
 
@@ -18,14 +19,9 @@ use tower_mcp::{McpRouter, StdioTransport};
 
 use skillet_mcp::cache::{self, RegistrySource};
 use skillet_mcp::config;
-use skillet_mcp::install::{self, InstallOptions};
-use skillet_mcp::manifest;
 use skillet_mcp::registry::{cache_dir_for_url, default_cache_dir, parse_duration};
 use skillet_mcp::state::AppState;
-use skillet_mcp::{
-    discover, git, index, integrity, pack, publish, registry, safety, scaffold, search, state,
-    trust, validate,
-};
+use skillet_mcp::{discover, git, index, registry, search, state};
 
 #[derive(Parser, Debug)]
 #[command(name = "skillet")]
@@ -377,19 +373,19 @@ async fn main() -> ExitCode {
     let is_serve = matches!(cli.command, Some(Command::Serve(_)) | None);
 
     let exit_code = match cli.command {
-        Some(Command::Validate(args)) => run_validate(args),
-        Some(Command::Pack(args)) => run_pack(args),
-        Some(Command::Publish(args)) => run_publish(args),
-        Some(Command::InitRegistry(args)) => run_init_registry(args),
-        Some(Command::InitSkill(args)) => run_init_skill(args),
-        Some(Command::Install(args)) => run_install(args),
-        Some(Command::Search(args)) => run_search(args),
-        Some(Command::Categories(args)) => run_categories(args),
-        Some(Command::Info(args)) => run_info(args),
-        Some(Command::List(args)) => run_list(args),
-        Some(Command::Trust(args)) => run_trust(args),
-        Some(Command::Audit(args)) => run_audit(args),
-        Some(Command::Setup(args)) => run_setup(args),
+        Some(Command::Validate(args)) => cli::author::run_validate(args),
+        Some(Command::Pack(args)) => cli::author::run_pack(args),
+        Some(Command::Publish(args)) => cli::author::run_publish(args),
+        Some(Command::InitRegistry(args)) => cli::author::run_init_registry(args),
+        Some(Command::InitSkill(args)) => cli::author::run_init_skill(args),
+        Some(Command::Install(args)) => cli::install::run_install(args),
+        Some(Command::Search(args)) => cli::search::run_search(args),
+        Some(Command::Categories(args)) => cli::search::run_categories(args),
+        Some(Command::Info(args)) => cli::search::run_info(args),
+        Some(Command::List(args)) => cli::search::run_list(args),
+        Some(Command::Trust(args)) => cli::trust::run_trust(args),
+        Some(Command::Audit(args)) => cli::trust::run_audit(args),
+        Some(Command::Setup(args)) => cli::setup::run_setup(args),
         Some(Command::Serve(args)) => run_serve(args).await,
         None => run_serve(cli.serve).await,
     };
@@ -400,1279 +396,6 @@ async fn main() -> ExitCode {
     }
 
     exit_code
-}
-
-/// Run the `validate` subcommand.
-fn run_validate(args: ValidateArgs) -> ExitCode {
-    let path = &args.path;
-    println!("Validating {} ...\n", path.display());
-
-    let result = match validate::validate_skillpack(path) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("  error: {e}");
-            eprintln!("\nValidation failed.");
-            return ExitCode::from(1);
-        }
-    };
-
-    // skill.toml
-    println!("  skill.toml ............ ok");
-
-    // SKILL.md
-    let line_count = result.skill_md.lines().count();
-    println!("  SKILL.md .............. ok ({line_count} lines)");
-
-    // Core fields
-    println!("  owner ................. {}", result.owner);
-    println!("  name .................. {}", result.name);
-    println!("  version ............... {}", result.version);
-    println!("  description ........... {}", result.description);
-
-    // Categories and tags
-    if let Some(ref classification) = result.metadata.skill.classification {
-        if !classification.categories.is_empty() {
-            println!(
-                "  categories ............ {}",
-                classification.categories.join(", ")
-            );
-        }
-        if !classification.tags.is_empty() {
-            println!(
-                "  tags .................. {}",
-                classification.tags.join(", ")
-            );
-        }
-    }
-
-    // Extra files
-    if !result.files.is_empty() {
-        let mut file_paths: Vec<&String> = result.files.keys().collect();
-        file_paths.sort();
-        println!(
-            "  extra files ........... {}",
-            file_paths
-                .iter()
-                .map(|s| s.as_str())
-                .collect::<Vec<_>>()
-                .join(", ")
-        );
-    }
-
-    // Content hash (show abbreviated)
-    let hash_display = if result.hashes.composite.len() > 17 {
-        format!("{}...", &result.hashes.composite[..17])
-    } else {
-        result.hashes.composite.clone()
-    };
-    println!("  content hash .......... {hash_display}");
-
-    // Manifest status
-    match result.manifest_ok {
-        Some(true) => println!("  manifest .............. verified"),
-        Some(false) => println!("  manifest .............. MISMATCH"),
-        None => println!("  manifest .............. not found (will be generated on publish)"),
-    }
-
-    // Warnings
-    if !result.warnings.is_empty() {
-        println!();
-        for w in &result.warnings {
-            println!("  warning: {w}");
-        }
-    }
-
-    // Safety scanning
-    if !args.skip_safety {
-        let cli_config = config::load_config().unwrap_or_default();
-        let report = safety::scan(
-            &result.skill_md,
-            &result.skill_toml_raw,
-            &result.files,
-            &result.metadata,
-            &cli_config.safety.suppress,
-        );
-
-        if !report.is_empty() {
-            println!();
-            print_safety_report(&report);
-        }
-
-        if report.has_danger() {
-            eprintln!("\nValidation failed: safety issues detected.");
-            return ExitCode::from(2);
-        }
-    }
-
-    println!("\nValidation passed.");
-    ExitCode::SUCCESS
-}
-
-/// Run the `pack` subcommand.
-fn run_pack(args: PackArgs) -> ExitCode {
-    let path = &args.path;
-    println!("Packing {} ...\n", path.display());
-
-    let result = match pack::pack(path) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("  error: {e}");
-            eprintln!("\nPack failed.");
-            return ExitCode::from(1);
-        }
-    };
-
-    let v = &result.validation;
-    println!("  owner ................. {}", v.owner);
-    println!("  name .................. {}", v.name);
-    println!("  version ............... {}", v.version);
-
-    if result.manifest_written {
-        println!("  MANIFEST.sha256 ....... written");
-    }
-
-    if result.versions_updated {
-        println!("  versions.toml ......... updated");
-    } else {
-        println!("  versions.toml ......... up to date");
-    }
-
-    // Safety scanning
-    if !args.skip_safety {
-        let cli_config = config::load_config().unwrap_or_default();
-        let report = safety::scan(
-            &v.skill_md,
-            &v.skill_toml_raw,
-            &v.files,
-            &v.metadata,
-            &cli_config.safety.suppress,
-        );
-
-        if !report.is_empty() {
-            println!();
-            print_safety_report(&report);
-        }
-
-        if report.has_danger() {
-            eprintln!("\nPack failed: safety issues detected.");
-            return ExitCode::from(2);
-        }
-    }
-
-    println!("\nPack succeeded.");
-    ExitCode::SUCCESS
-}
-
-/// Run the `publish` subcommand.
-fn run_publish(args: PublishArgs) -> ExitCode {
-    let path = &args.path;
-    println!("Publishing {} to {} ...\n", path.display(), args.repo);
-
-    let result = match publish::publish(path, &args.repo, args.dry_run) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("  error: {e}");
-            eprintln!("\nPublish failed.");
-            return ExitCode::from(1);
-        }
-    };
-
-    let v = &result.pack.validation;
-    println!("  owner ................. {}", v.owner);
-    println!("  name .................. {}", v.name);
-    println!("  version ............... {}", v.version);
-
-    // Safety scanning
-    if !args.skip_safety {
-        let cli_config = config::load_config().unwrap_or_default();
-        let report = safety::scan(
-            &v.skill_md,
-            &v.skill_toml_raw,
-            &v.files,
-            &v.metadata,
-            &cli_config.safety.suppress,
-        );
-
-        if !report.is_empty() {
-            println!();
-            print_safety_report(&report);
-        }
-
-        if report.has_danger() {
-            eprintln!("\nPublish failed: safety issues detected.");
-            return ExitCode::from(2);
-        }
-    }
-
-    if args.dry_run {
-        println!("\nDry run complete.");
-    } else {
-        println!("  PR .................... {}", result.pr_url);
-        println!("\nPublish succeeded.");
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `init-registry` subcommand.
-fn run_init_registry(args: InitRegistryArgs) -> ExitCode {
-    let path = &args.path;
-
-    if path.exists() {
-        eprintln!("Error: {} already exists", path.display());
-        return ExitCode::from(1);
-    }
-
-    let name = args
-        .name
-        .or_else(|| {
-            path.file_name()
-                .and_then(|n| n.to_str())
-                .map(|s| s.to_string())
-        })
-        .unwrap_or_else(|| "my-skills".to_string());
-
-    if let Err(e) = registry::init_registry(path, &name, args.description.as_deref()) {
-        eprintln!("Error: {e}");
-        return ExitCode::from(1);
-    }
-
-    println!("Initialized skill registry at {}", path.display());
-    println!();
-    println!("  cd {}", path.display());
-    println!("  # add skills: mkdir -p owner/skill-name");
-    println!("  # serve locally: skillet --registry .");
-    println!("  # push and serve remotely: skillet --remote <git-url>");
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `init-skill` subcommand.
-fn run_init_skill(args: InitSkillArgs) -> ExitCode {
-    let path = &args.path;
-
-    // Infer owner/name from path components
-    let name = match path.file_name().and_then(|n| n.to_str()) {
-        Some(n) if !n.is_empty() => n.to_string(),
-        _ => {
-            eprintln!("Error: could not infer skill name from path");
-            return ExitCode::from(1);
-        }
-    };
-
-    let owner = match path
-        .parent()
-        .and_then(|p| p.file_name())
-        .and_then(|n| n.to_str())
-    {
-        Some(o) if !o.is_empty() => o.to_string(),
-        _ => {
-            eprintln!(
-                "Error: could not infer owner from path. Use owner/skill-name format (e.g. myname/my-skill)"
-            );
-            return ExitCode::from(1);
-        }
-    };
-
-    let description = args
-        .description
-        .unwrap_or_else(|| format!("A skill for {name}"));
-
-    if let Err(e) = scaffold::init_skill(
-        path,
-        &owner,
-        &name,
-        &description,
-        &args.category,
-        &args.tags,
-    ) {
-        eprintln!("Error: {e}");
-        return ExitCode::from(1);
-    }
-
-    println!("Created skillpack at {}", path.display());
-    println!();
-    println!("  owner ................. {owner}");
-    println!("  name .................. {name}");
-    println!();
-    println!("Next steps:");
-    println!(
-        "  1. Edit {}/skill.toml to customize metadata",
-        path.display()
-    );
-    println!(
-        "  2. Edit {}/SKILL.md to write your skill prompt",
-        path.display()
-    );
-    println!("  3. Validate: skillet validate {}", path.display());
-
-    ExitCode::SUCCESS
-}
-
-/// Parse an "owner/name" skill reference.
-fn parse_skill_ref(s: &str) -> Result<(&str, &str), String> {
-    let parts: Vec<&str> = s.splitn(2, '/').collect();
-    if parts.len() != 2 || parts[0].is_empty() || parts[1].is_empty() {
-        return Err(format!(
-            "Invalid skill reference: '{s}'. Expected format: owner/name"
-        ));
-    }
-    Ok((parts[0], parts[1]))
-}
-
-/// Run the `install` subcommand.
-fn run_install(args: InstallArgs) -> ExitCode {
-    let (owner, name) = match parse_skill_ref(&args.skill) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let mut cli_config = match config::load_config() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error loading config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if args.registries.no_cache {
-        cli_config.cache.enabled = false;
-    }
-
-    let targets = match config::resolve_targets(&args.target, &cli_config) {
-        Ok(t) => t,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let global = args.global || cli_config.install.global;
-
-    let (skill_index, registry_paths) = match registry::load_registries(
-        &args.registries.registry,
-        &args.registries.remote,
-        &cli_config,
-        args.registries.subdir.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error loading registries: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    // Look up the skill
-    let entry = match skill_index
-        .skills
-        .get(&(owner.to_string(), name.to_string()))
-    {
-        Some(e) => e,
-        None => {
-            eprintln!("Error: skill '{owner}/{name}' not found in any registry");
-            return ExitCode::from(1);
-        }
-    };
-
-    // Resolve version
-    let version = if let Some(ref requested) = args.version {
-        match entry.versions.iter().find(|v| v.version == *requested) {
-            Some(v) if !v.has_content => {
-                eprintln!(
-                    "Error: version {requested} exists but content is not available \
-                     (only the latest version has full content)"
-                );
-                return ExitCode::from(1);
-            }
-            Some(v) => v,
-            None => {
-                let available: Vec<&str> =
-                    entry.versions.iter().map(|v| v.version.as_str()).collect();
-                eprintln!(
-                    "Error: version '{requested}' not found for {owner}/{name}\n\
-                     Available versions: {}",
-                    available.join(", ")
-                );
-                return ExitCode::from(1);
-            }
-        }
-    } else {
-        match entry.latest() {
-            Some(v) => v,
-            None => {
-                eprintln!("Error: no available versions for {owner}/{name} (all yanked)");
-                return ExitCode::from(1);
-            }
-        }
-    };
-
-    // Load manifest
-    let mut installed_manifest = match manifest::load() {
-        Ok(m) => m,
-        Err(e) => {
-            eprintln!("Error loading installation manifest: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    // Determine registry identifier
-    let registry_id = if !registry_paths.is_empty() {
-        registry::registry_id(&registry_paths[0], &args.registries.remote)
-    } else {
-        "unknown".to_string()
-    };
-
-    // Trust checking
-    let content_hash = integrity::sha256_hex(&version.skill_md);
-    let trust_state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let trust_check = trust::check_trust(&trust_state, &registry_id, owner, name, &content_hash);
-
-    match trust_check.tier {
-        trust::TrustTier::Trusted => {}
-        trust::TrustTier::Reviewed => {
-            if trust_check.pinned_hash.as_deref() != Some(&content_hash) {
-                eprintln!(
-                    "Warning: {} (tier: reviewed, content changed since pinned)",
-                    trust_check.reason
-                );
-            }
-        }
-        trust::TrustTier::Unknown => {
-            // --require-trusted flag or config overrides policy
-            if args.require_trusted || cli_config.trust.require_trusted {
-                eprintln!(
-                    "Error: {reason}\n\n\
-                     Install blocked: --require-trusted is set.\n\
-                     To install this skill, either:\n\
-                     \n  1. Trust the registry:\n\
-                     \n     skillet trust add-registry {registry_id}\n\
-                     \n  2. Review and pin the skill:\n\
-                     \n     skillet info {owner}/{name}\n\
-                     \n     skillet trust pin {owner}/{name}\n\
-                     \n     skillet install {owner}/{name}\n",
-                    reason = trust_check.reason,
-                );
-                return ExitCode::from(1);
-            }
-
-            let policy = &cli_config.trust.unknown_policy;
-            match policy.as_str() {
-                "block" => {
-                    eprintln!(
-                        "Error: {reason}\n\n\
-                         Install blocked by trust policy (unknown_policy = \"block\").\n\
-                         To install this skill, either:\n\
-                         \n  1. Trust the registry:\n\
-                         \n     skillet trust add-registry {registry_id}\n\
-                         \n  2. Review and pin the skill:\n\
-                         \n     skillet info {owner}/{name}\n\
-                         \n     skillet trust pin {owner}/{name}\n\
-                         \n     skillet install {owner}/{name}\n",
-                        reason = trust_check.reason,
-                    );
-                    return ExitCode::from(1);
-                }
-                "prompt" => {
-                    eprintln!(
-                        "Warning: {}\nProceed with install? [y/N] ",
-                        trust_check.reason
-                    );
-                    let mut input = String::new();
-                    if std::io::stdin().read_line(&mut input).is_err()
-                        || !input.trim().eq_ignore_ascii_case("y")
-                    {
-                        eprintln!("Install cancelled.");
-                        return ExitCode::from(1);
-                    }
-                }
-                _ => {
-                    // "warn" (default) -- explicit guidance
-                    eprintln!(
-                        "Warning: {reason}\n\
-                         To verify before installing:\n\
-                         \n  skillet info {owner}/{name}\n\
-                         \n  skillet trust pin {owner}/{name}\n",
-                        reason = trust_check.reason,
-                    );
-                }
-            }
-        }
-    }
-
-    let options = InstallOptions {
-        targets,
-        global,
-        registry: registry_id.clone(),
-    };
-
-    let results =
-        match install::install_skill(owner, name, version, &options, &mut installed_manifest) {
-            Ok(r) => r,
-            Err(e) => {
-                eprintln!("Error installing skill: {e}");
-                return ExitCode::from(1);
-            }
-        };
-
-    // Save manifest
-    if let Err(e) = manifest::save(&installed_manifest) {
-        eprintln!("Error saving installation manifest: {e}");
-        return ExitCode::from(1);
-    }
-
-    // Auto-pin content hash after successful install
-    if cli_config.trust.auto_pin {
-        let mut trust_state = trust_state;
-        trust_state.pin_skill(owner, name, &version.version, &registry_id, &content_hash);
-        if let Err(e) = trust::save(&trust_state) {
-            eprintln!("Warning: failed to save trust state: {e}");
-        }
-    }
-
-    // Print results
-    println!(
-        "Installed {owner}/{name} v{version}",
-        version = version.version
-    );
-    println!();
-    for r in &results {
-        let file_count = r.files_written.len();
-        let scope = if options.global { "global" } else { "project" };
-        println!(
-            "  {target} ({scope}) ... {file_count} file{s} -> {path}",
-            target = r.target,
-            s = if file_count == 1 { "" } else { "s" },
-            path = r.path.display(),
-        );
-    }
-
-    // Safety scanning (informational only -- never blocks install)
-    let report = safety::scan(
-        &version.skill_md,
-        &version.skill_toml_raw,
-        &version.files,
-        &version.metadata,
-        &cli_config.safety.suppress,
-    );
-
-    if !report.is_empty() {
-        println!();
-        print_safety_report(&report);
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `categories` subcommand.
-fn run_categories(args: CategoriesArgs) -> ExitCode {
-    let mut cli_config = match config::load_config() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error loading config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if args.registries.no_cache {
-        cli_config.cache.enabled = false;
-    }
-
-    let (skill_index, _registry_paths) = match registry::load_registries(
-        &args.registries.registry,
-        &args.registries.remote,
-        &cli_config,
-        args.registries.subdir.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error loading registries: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if skill_index.categories.is_empty() {
-        println!("No categories found.");
-        return ExitCode::SUCCESS;
-    }
-
-    let total: usize = skill_index.categories.values().sum();
-    println!(
-        "{} categor{} ({total} skill{}):\n",
-        skill_index.categories.len(),
-        if skill_index.categories.len() == 1 {
-            "y"
-        } else {
-            "ies"
-        },
-        if total == 1 { "" } else { "s" },
-    );
-    for (name, count) in &skill_index.categories {
-        println!("  {name} ({count})");
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `search` subcommand.
-fn run_search(args: SearchArgs) -> ExitCode {
-    let mut cli_config = match config::load_config() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error loading config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if args.registries.no_cache {
-        cli_config.cache.enabled = false;
-    }
-
-    let (skill_index, _registry_paths) = match registry::load_registries(
-        &args.registries.registry,
-        &args.registries.remote,
-        &cli_config,
-        args.registries.subdir.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error loading registries: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let skill_search = search::SkillSearch::build(&skill_index);
-
-    // Wildcard: list all skills
-    let results: Vec<state::SkillSummary> = if args.query == "*" {
-        let mut keys: Vec<_> = skill_index.skills.keys().collect();
-        keys.sort();
-        keys.iter()
-            .filter_map(|k| {
-                let entry = skill_index.skills.get(*k)?;
-                state::SkillSummary::from_entry(entry)
-            })
-            .collect()
-    } else {
-        let hits = skill_search.search(&args.query, 20);
-        hits.iter()
-            .filter_map(|(owner, name, _score)| {
-                let entry = skill_index.skills.get(&(owner.clone(), name.clone()))?;
-                state::SkillSummary::from_entry(entry)
-            })
-            .collect()
-    };
-
-    // Apply structured filters
-    let results: Vec<_> = results
-        .into_iter()
-        .filter(|s| {
-            if let Some(ref cat) = args.category
-                && !s.categories.iter().any(|c| c.eq_ignore_ascii_case(cat))
-            {
-                return false;
-            }
-            if let Some(ref tag) = args.tag
-                && !s.tags.iter().any(|t| t.eq_ignore_ascii_case(tag))
-            {
-                return false;
-            }
-            if let Some(ref owner) = args.owner
-                && !s.owner.eq_ignore_ascii_case(owner)
-            {
-                return false;
-            }
-            true
-        })
-        .collect();
-
-    if results.is_empty() {
-        println!("No skills found.");
-        return ExitCode::SUCCESS;
-    }
-
-    println!(
-        "Found {} skill{}:\n",
-        results.len(),
-        if results.len() == 1 { "" } else { "s" }
-    );
-    for s in &results {
-        println!("  {}/{} v{}", s.owner, s.name, s.version);
-        println!("    {}", s.description);
-        if !s.categories.is_empty() {
-            println!("    categories: {}", s.categories.join(", "));
-        }
-        if !s.tags.is_empty() {
-            println!("    tags: {}", s.tags.join(", "));
-        }
-        println!();
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `info` subcommand.
-fn run_info(args: InfoArgs) -> ExitCode {
-    let (owner, name) = match parse_skill_ref(&args.skill) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let mut cli_config = match config::load_config() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error loading config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if args.registries.no_cache {
-        cli_config.cache.enabled = false;
-    }
-
-    let (skill_index, _registry_paths) = match registry::load_registries(
-        &args.registries.registry,
-        &args.registries.remote,
-        &cli_config,
-        args.registries.subdir.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error loading registries: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let entry = match skill_index
-        .skills
-        .get(&(owner.to_string(), name.to_string()))
-    {
-        Some(e) => e,
-        None => {
-            eprintln!("Error: skill '{owner}/{name}' not found in any registry");
-            return ExitCode::from(1);
-        }
-    };
-
-    let latest = match entry.latest() {
-        Some(v) => v,
-        None => {
-            eprintln!("Error: no available versions for {owner}/{name} (all yanked)");
-            return ExitCode::from(1);
-        }
-    };
-
-    let info = &latest.metadata.skill;
-
-    println!("{owner}/{name}\n");
-    println!("  version ............... {}", info.version);
-    println!("  description ........... {}", info.description);
-
-    if let Some(ref trigger) = info.trigger {
-        println!("  trigger ............... {trigger}");
-    }
-    if let Some(ref license) = info.license {
-        println!("  license ............... {license}");
-    }
-    if let Some(ref author) = info.author {
-        if let Some(ref name) = author.name {
-            println!("  author ................ {name}");
-        }
-        if let Some(ref github) = author.github {
-            println!("  github ................ {github}");
-        }
-    }
-    if let Some(ref classification) = info.classification {
-        if !classification.categories.is_empty() {
-            println!(
-                "  categories ............ {}",
-                classification.categories.join(", ")
-            );
-        }
-        if !classification.tags.is_empty() {
-            println!(
-                "  tags .................. {}",
-                classification.tags.join(", ")
-            );
-        }
-    }
-    if let Some(ref compat) = info.compatibility
-        && !compat.verified_with.is_empty()
-    {
-        println!(
-            "  verified with ......... {}",
-            compat.verified_with.join(", ")
-        );
-    }
-
-    // Extra files
-    if !latest.files.is_empty() {
-        let mut file_paths: Vec<&String> = latest.files.keys().collect();
-        file_paths.sort();
-        println!(
-            "  files ................. {}",
-            file_paths
-                .iter()
-                .map(|s| s.as_str())
-                .collect::<Vec<_>>()
-                .join(", ")
-        );
-    }
-
-    // Published timestamp
-    if let Some(ref published) = latest.published {
-        println!("  published ............. {published}");
-    }
-
-    // Content hash
-    if let Some(ref hash) = latest.content_hash {
-        let display = if hash.len() > 17 {
-            format!("{}...", &hash[..17])
-        } else {
-            hash.clone()
-        };
-        println!("  content hash .......... {display}");
-    }
-
-    // Integrity
-    match latest.integrity_ok {
-        Some(true) => println!("  integrity ............. verified"),
-        Some(false) => println!("  integrity ............. MISMATCH"),
-        None => {}
-    }
-
-    // Version history
-    let available: Vec<&str> = entry
-        .versions
-        .iter()
-        .filter(|v| !v.yanked)
-        .map(|v| v.version.as_str())
-        .collect();
-    if available.len() > 1 {
-        println!("  versions .............. {}", available.join(", "));
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `list` subcommand.
-fn run_list(_args: ListArgs) -> ExitCode {
-    let installed_manifest = match manifest::load() {
-        Ok(m) => m,
-        Err(e) => {
-            eprintln!("Error loading installation manifest: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if installed_manifest.skills.is_empty() {
-        println!("No skills installed.");
-        return ExitCode::SUCCESS;
-    }
-
-    // Group by (owner, name)
-    let mut groups: std::collections::BTreeMap<(String, String), Vec<&manifest::InstalledSkill>> =
-        std::collections::BTreeMap::new();
-    for skill in &installed_manifest.skills {
-        groups
-            .entry((skill.owner.clone(), skill.name.clone()))
-            .or_default()
-            .push(skill);
-    }
-
-    println!(
-        "{} skill{} installed:\n",
-        groups.len(),
-        if groups.len() == 1 { "" } else { "s" }
-    );
-
-    for ((owner, name), entries) in &groups {
-        let version = &entries[0].version;
-        println!("  {owner}/{name} v{version}");
-        for entry in entries {
-            println!(
-                "    -> {}  ({})",
-                entry.installed_to.display(),
-                entry.installed_at,
-            );
-        }
-        println!();
-    }
-
-    ExitCode::SUCCESS
-}
-
-/// Run the `trust` subcommand.
-fn run_trust(args: TrustArgs) -> ExitCode {
-    match args.action {
-        TrustAction::AddRegistry(a) => run_trust_add_registry(a),
-        TrustAction::RemoveRegistry(a) => run_trust_remove_registry(a),
-        TrustAction::List(a) => run_trust_list(a),
-        TrustAction::Pin(a) => run_trust_pin(a),
-        TrustAction::Unpin(a) => run_trust_unpin(a),
-    }
-}
-
-fn run_trust_add_registry(args: TrustAddRegistryArgs) -> ExitCode {
-    let mut state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if state.is_trusted(&args.url) {
-        println!("Registry already trusted: {}", args.url);
-        return ExitCode::SUCCESS;
-    }
-
-    state.add_registry(&args.url, args.note.as_deref());
-
-    if let Err(e) = trust::save(&state) {
-        eprintln!("Error saving trust state: {e}");
-        return ExitCode::from(1);
-    }
-
-    println!("Trusted: {}", args.url);
-    ExitCode::SUCCESS
-}
-
-fn run_trust_remove_registry(args: TrustRemoveRegistryArgs) -> ExitCode {
-    let mut state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if !state.remove_registry(&args.url) {
-        eprintln!("Registry not found in trusted list: {}", args.url);
-        return ExitCode::from(1);
-    }
-
-    if let Err(e) = trust::save(&state) {
-        eprintln!("Error saving trust state: {e}");
-        return ExitCode::from(1);
-    }
-
-    println!("Removed: {}", args.url);
-    ExitCode::SUCCESS
-}
-
-fn run_trust_list(args: TrustListArgs) -> ExitCode {
-    let state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if state.trusted_registries.is_empty() && state.pinned_skills.is_empty() {
-        println!("No trusted registries or pinned skills.");
-        return ExitCode::SUCCESS;
-    }
-
-    if !state.trusted_registries.is_empty() {
-        println!("Trusted registries ({}):\n", state.trusted_registries.len());
-        for r in &state.trusted_registries {
-            print!("  {}", r.registry);
-            if let Some(ref note) = r.note {
-                print!("  ({note})");
-            }
-            println!("  [{}]", r.trusted_at);
-        }
-    }
-
-    if !args.registries_only && !state.pinned_skills.is_empty() {
-        if !state.trusted_registries.is_empty() {
-            println!();
-        }
-        println!("Pinned skills ({}):\n", state.pinned_skills.len());
-        for p in &state.pinned_skills {
-            let hash_display = if p.content_hash.len() > 17 {
-                format!("{}...", &p.content_hash[..17])
-            } else {
-                p.content_hash.clone()
-            };
-            println!(
-                "  {}/{} v{}  {}  [{}]",
-                p.owner, p.name, p.version, hash_display, p.pinned_at
-            );
-        }
-    }
-
-    ExitCode::SUCCESS
-}
-
-fn run_trust_pin(args: TrustPinArgs) -> ExitCode {
-    let (owner, name) = match parse_skill_ref(&args.skill) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let mut cli_config = match config::load_config() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error loading config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if args.registries.no_cache {
-        cli_config.cache.enabled = false;
-    }
-
-    let (skill_index, registry_paths) = match registry::load_registries(
-        &args.registries.registry,
-        &args.registries.remote,
-        &cli_config,
-        args.registries.subdir.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error loading registries: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let entry = match skill_index
-        .skills
-        .get(&(owner.to_string(), name.to_string()))
-    {
-        Some(e) => e,
-        None => {
-            eprintln!("Error: skill '{owner}/{name}' not found in any registry");
-            return ExitCode::from(1);
-        }
-    };
-
-    let version = match entry.latest() {
-        Some(v) => v,
-        None => {
-            eprintln!("Error: no available versions for {owner}/{name} (all yanked)");
-            return ExitCode::from(1);
-        }
-    };
-
-    let registry_id = if !registry_paths.is_empty() {
-        registry::registry_id(&registry_paths[0], &args.registries.remote)
-    } else {
-        "unknown".to_string()
-    };
-
-    let content_hash = integrity::sha256_hex(&version.skill_md);
-
-    let mut trust_state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    trust_state.pin_skill(owner, name, &version.version, &registry_id, &content_hash);
-
-    if let Err(e) = trust::save(&trust_state) {
-        eprintln!("Error saving trust state: {e}");
-        return ExitCode::from(1);
-    }
-
-    let hash_display = if content_hash.len() > 17 {
-        format!("{}...", &content_hash[..17])
-    } else {
-        content_hash.clone()
-    };
-    println!(
-        "Pinned {owner}/{name} v{} ({hash_display})",
-        version.version
-    );
-    ExitCode::SUCCESS
-}
-
-fn run_trust_unpin(args: TrustUnpinArgs) -> ExitCode {
-    let (owner, name) = match parse_skill_ref(&args.skill) {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let mut state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    if !state.unpin_skill(owner, name) {
-        eprintln!("Skill '{owner}/{name}' is not pinned.");
-        return ExitCode::from(1);
-    }
-
-    if let Err(e) = trust::save(&state) {
-        eprintln!("Error saving trust state: {e}");
-        return ExitCode::from(1);
-    }
-
-    println!("Unpinned {owner}/{name}");
-    ExitCode::SUCCESS
-}
-
-/// Run the `audit` subcommand.
-fn run_audit(args: AuditArgs) -> ExitCode {
-    let installed = match manifest::load() {
-        Ok(m) => m,
-        Err(e) => {
-            eprintln!("Error loading installation manifest: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let trust_state = match trust::load() {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error loading trust state: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let (filter_owner, filter_name) = if let Some(ref skill) = args.skill {
-        match parse_skill_ref(skill) {
-            Ok((o, n)) => (Some(o), Some(n)),
-            Err(e) => {
-                eprintln!("Error: {e}");
-                return ExitCode::from(1);
-            }
-        }
-    } else {
-        (None, None)
-    };
-
-    let results = trust::audit(&installed, &trust_state, filter_owner, filter_name);
-
-    if results.is_empty() {
-        println!("No installed skills to audit.");
-        return ExitCode::SUCCESS;
-    }
-
-    for r in &results {
-        println!(
-            "  [{status}] {owner}/{name} v{version} -> {path}",
-            status = r.status,
-            owner = r.owner,
-            name = r.name,
-            version = r.version,
-            path = r.installed_to.display(),
-        );
-    }
-
-    if trust::audit_has_problems(&results) {
-        ExitCode::from(1)
-    } else {
-        ExitCode::SUCCESS
-    }
-}
-
-/// Run the `setup` subcommand.
-fn run_setup(args: SetupArgs) -> ExitCode {
-    let config_path = config::config_dir().join("config.toml");
-
-    // Check for existing config
-    if config_path.exists() && !args.force {
-        eprintln!(
-            "Config already exists at {}\nUse --force to overwrite.",
-            config_path.display()
-        );
-        return ExitCode::from(1);
-    }
-
-    // Build remotes list: official first (unless opted out), then user-provided
-    let mut remotes = Vec::new();
-    if !args.no_official_registry {
-        remotes.push(registry::DEFAULT_REGISTRY_URL.to_string());
-    }
-    remotes.extend(args.remote);
-
-    let config = match config::generate_default_config(remotes, args.registry, &args.target) {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    let path = match config::write_config(&config) {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("Error writing config: {e}");
-            return ExitCode::from(1);
-        }
-    };
-
-    // Read back the written content to display
-    let content = std::fs::read_to_string(&path).unwrap_or_default();
-
-    println!("Wrote {}\n", path.display());
-    println!("{content}");
-    println!(
-        "To use skillet with your agent, add this to your MCP config:\n\n\
-         {{\n  \"mcpServers\": {{\n    \"skillet\": {{\n      \"command\": \"skillet\"\n    }}\n  }}\n}}"
-    );
-    println!(
-        "\nNext steps:\n  \
-         skillet search *            # browse available skills\n  \
-         skillet info owner/name     # see skill details\n  \
-         skillet install owner/name  # install a skill"
-    );
-
-    ExitCode::SUCCESS
-}
-
-/// Print a safety report to stdout/stderr.
-fn print_safety_report(report: &safety::SafetyReport) {
-    let danger_count = report
-        .findings
-        .iter()
-        .filter(|f| f.severity == safety::Severity::Danger)
-        .count();
-    let warning_count = report
-        .findings
-        .iter()
-        .filter(|f| f.severity == safety::Severity::Warning)
-        .count();
-
-    println!("Safety scan: {danger_count} danger, {warning_count} warning\n");
-
-    for f in &report.findings {
-        let line_info = match f.line {
-            Some(n) => format!("{}:{n}", f.file),
-            None => f.file.clone(),
-        };
-        println!("  [{severity}] {line_info}", severity = f.severity);
-        println!("    rule: {}", f.rule_id);
-        println!("    {}", f.message);
-        println!("    matched: {}", f.matched);
-    }
 }
 
 /// All known tool short names.
